requirements.txt
Django==5.0.6
djangorestframework==3.15.1
drf-spectacular==0.27.2
mysqlclient==2.2.4
django-cors-headers==4.3.1